Move rule-provider-executions to executions module.
